Fixes and an enhancement
The AdvancedRoute package was very helpful when working on our port to
Laravel 5.3, so many thanks for the work in creating it. We have over
700 routes across more than 30 controllers, so trying to replace the
deprecated Route::controller manually would have been a nightmare.
However, we did encounter a couple of issues in this package which I
had to fix and I also added one enhancement that was really useful to
us.

Firstly, the slug method was using a str_replace to remove the http
method name from the start of controller method, but unfortuately that
removed all occurances of all the http method names from the entire
method name. So example, the method getCompanyname, became "compname"
instead of "companyname". This method now uses a preg to anchor the
search to the start of the method name.

Secondly, in addition to get, post and any, we also use the patch and
delete http methods, so I generalized the code to support both of those
and, for completness, "put" as well.

Lastly, we wanted to move away from ::controller routing entirely and
use explicit route commands, but as I mentioned above that would have
mean crafting 700+ statements for the routes/web.php file. However,
since AdvancedRoute had already computed the necessary routes, I added
the "optional" capability to have it emit all those routes in the exact
format required for insertion into routes/web.php. If the
EMIT_ROUTE_STATEMENTS class variable is set to true, a
/tmp/controllerRoutes directory will be created and one file for each
call to AdvancedRoute::controller will be written to that directory
containing the computed routes. Obviously this only needs to be done
once.